Day 4/task4/pyproject.toml
[project]
name = "task4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.9.2",
    "openai-agents>=0.6.1",
    "pypdf>=6.3.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
